requirements.txt
kivy>=2.3.0
kivymd>=1.2.0
rns>=1.8.0
lxmf>=0.9.0
pillow>=10.0.0
msgpack>=1.0.5
requests>=2.31.0
cryptography>=41.0.0
plyer>=2.1.0
